% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qar1.R
\name{Q.ar1}
\alias{Q.ar1}
\title{Precision matrix for an AR1 process}
\usage{
Q.ar1(N, sigma, rho, vcov = FALSE)
}
\arguments{
\item{N}{int > 0, number of obserations you want to observe}

\item{sigma}{float > 0, pairwise observation variance}

\item{rho}{float >= 0 & < 1, how correlated pairwise observations are. The
function will still run with values outside of the range [0,1) however the
stability of the simulation results are not gaurunteed.}

\item{vcov}{bool If the vcov matrix should be returned instead of the
precision matrix.}
}
\value{
Matrix object, either precision or variance covariance
}
\description{
Creates a precision (or optionally variance-covariance) matrix
of an AR1 process
}

