% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QlCAR.R
\name{Q.lCAR}
\alias{Q.lCAR}
\title{Precision matrix for a pCAR process}
\usage{
Q.lCAR(graph, sigma, rho, vcov = FALSE)
}
\arguments{
\item{graph}{matrix, square matrix indicating where two observations are
connected (and therefore conditionally auto-regressive).}

\item{sigma}{float > 0, pairwise observation variance}

\item{rho}{float >= 0 & < 1, how correlated pairwise observations are. The
function will still run with values outside of the range [0,1) however the
stability of the simulation results are not gaurunteed.}

\item{vcov}{bool If the vcov matrix should be returned instead of the
precision matrix.}
}
\value{
Matrix object, either precision or variance covariance
}
\description{
Creates a precision (or optionally variance-covariance) matrix
of a Leroux CAR(lCAR) process as defined in MacNab 2011. The matrix defines
the precision of estimates when observations share connections which are
conditionally auto-regressive(CAR).
}

