% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q.RW1.R, R/r.RW1.R
\name{Q.RW1}
\alias{Q.RW1}
\alias{r.RW1}
\title{Precision matrix for an RW1 process}
\usage{
Q.RW1(M, sigma, sparse=TRUE)

r.RW1(n, M, sigma)
}
\arguments{
\item{M}{int > 0, number of elements in the RW1 process.}

\item{sigma}{float > 0, pairwise observation standard deviation.}

\item{sparse}{bool Should the matrix be of class 'dsCMatrix'}

\item{n}{int > 0, number of observations to simulate from the GMRF.}
}
\value{
Q.RW1 returns a precision matrix with a RW1 structure.

r.RW1 retrurns a matrix with n rows which are the n observations of an
Intrinsic Gaussian Markov random field  RW1 process.
}
\description{
Functions for creating precision matricies and observations of
an RW1 process
}
\examples{
require("ggplot2")
# simulate RW1 GMRF
obs <- r.RW1(100, M=30, sigma=1)
# resulting matrix is n x M
dim(obs)
# subtract off the first time point to more easily observe correlation
obs_adj <- obs - obs[,1]
# move objects to a data frame
rw1_df <- data.frame(obs=c(t(obs_adj)), realization=rep(1:100, each=30),
                     time=rep(1:30, 100))
# plot each realization
ggplot(data=rw1_df, aes(time, obs, group=realization, color=realization)) +
    geom_line()

}
