% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q.RW2.R, R/r.RW2.R
\name{Q.RW2}
\alias{Q.RW2}
\alias{r.RW2}
\title{Precision matrix for an RW2 process}
\usage{
Q.RW2(M, sigma, sparse=TRUE)

r.RW2(n, M, sigma)
}
\arguments{
\item{M}{int > 0, number of elements in the RW2 process.}

\item{sigma}{float > 0, pairwise observation standard deviation.}

\item{sparse}{bool Should the matrix be of class 'dsCMatrix'}

\item{n}{int > 0, number of observations to simulate from the GMRF.}
}
\value{
Q.RW2 returns a precision matrix with a RW2 structure.

r.RW2 retrurns a matrix with n rows which are the n observations of an
Intrinsic Gaussian Markov random field  RW2 process.
}
\description{
Functions for creating precision matricies and observations of
an RW2 process
}
\examples{
require("ggplot2")
# simulate RW2 GMRF
obs <- r.RW2(100, M=30, sigma=1)
# resulting matrix is n x M
dim(obs)
# move objects to a data frame
RW2_df <- data.frame(obs=c(t(obs)), realization=rep(1:100, each=30),
                     time=rep(1:30, 100))
# plot each realization
ggplot(data=RW2_df, aes(time, obs, group=realization, color=realization)) +
    geom_line()

}
