% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rGMRF.R
\name{rGMRF}
\alias{rGMRF}
\title{Simulate data from a GMRF}
\usage{
rGMRF(n, Qstruct = c("Q.ar1", "Q.pCAR", "Q.lCAR"), ...)
}
\arguments{
\item{n}{int > 0, number of observations to simulate from the GMRF.}

\item{Qstruct}{string, indication of the type of covariance structure to use.}

\item{M}{int > 0, number of elements to observe, for AR1 process only}

\item{graph}{matrix, square matrix indicating where two observations are
connected (and therefore conditionally auto-regressive). Applicable to
lCAR & pCAR covariance structures.}

\item{sigma}{float > 0, pairwise observation variance}

\item{rho}{float >= 0 & < 1, how correlated pairwise observations are. The
function will still run with values outside of the range [0,1) however the
stability of the simulation results are not gaurunteed.}
}
\value{
Matrix object, matrix where each row is a single obsrevation from
a GMRF with covariance structure Q^-1.
}
\description{
Generlized function for simulating data from a Gaussian Markov
random field(GMRF. A string is passed in indicatting what the covariance
structure should be and then additional arguments are passed in based on the
requiremnts of the structure.
}

