% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simAR.R
\name{sim.AR}
\alias{sim.AR}
\title{Simulate correlated data from a precision matrix.}
\usage{
sim.AR(n, Q)
}
\arguments{
\item{n}{int > 0, number of observations to simulate from the GMRF.}

\item{Q}{matrix, a square precision matrix.}
}
\value{
Matrix object, matrix where each row is a single obsrevation from
a GMRF with covariance structure Q^-1.
}
\description{
Takes in a square precision matrix, which ideally should be
sparse and using Choleski factorization simulates data from a mean 0 process
where the inverse of the precision matrix represents the variance-covariance
of the points in the process. The resulting simulants represent samples of a
Gaussian Markov random field (GMRF).
}

