% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{US.df}
\alias{US.df}
\title{Spatial Polygons Data Frame of Counties for Several States}
\description{
Spatial Polygons data frame with 475 counties from the US states
Louisiana, Texas, Mississippi, & Arkansas. FIPS codes for the state and
county are provided in the data frame.
}
\keyword{data}
