% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{US.graph}
\alias{US.graph}
\title{Matrix of Shared Boundaries Between US.df Counties}
\description{
A 475x475 matrix where the index corresponds to a row in the
US.df Spatial Polygons data frame and the index of the matrix at row i column
j is 1 when US.df[i,] and US.df[j,] share a border and 0 when they do not.
}
\keyword{data}
