% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q.lCAR.R, R/r.lCAR.R
\name{Q.lCAR}
\alias{Q.lCAR}
\alias{r.lCAR}
\title{Precision matrix for a pCAR process}
\usage{
Q.lCAR(graph, sigma, rho, vcov=FALSE)

r.lCAR(n, graph, sigma, rho)
}
\arguments{
\item{graph}{matrix, square matrix indicating where two observations are
connected (and therefore conditionally auto-regressive).}

\item{sigma}{float > 0, pairwise observation variance}

\item{rho}{float >= 0 & < 1, how correlated pairwise observations are. The
function will still run with values outside of the range [0,1) however the
stability of the simulation results are not gaurunteed.}

\item{vcov}{bool If the vcov matrix should be returned instead of the
precision matrix.}

\item{n}{int > 0, number of observations to simulate from the GMRF.}
}
\value{
Q.lCAR returns either a precision or variance-covariance function
with a lCAR structure.

r.lCAR retrurns a matrix with n rows which are the n observations of a
Gaussian Markov random field lCAR process.
}
\description{
Functions for creating precision matricies and observations
of a Leroux CAR(lCAR) process as defined in MacNab 2011. The matrix defines
the precision of estimates when observations share connections which are
conditionally auto-regressive(CAR).
}

